execution
Figured out how to execute using a shell script.
Modified main class to deal with execution.

Compilation and execution can now be ran with:
./dam -f <pathname> -f <pathname>

Above execution now works but may need to be modified as 2 arguments
are neglected. Main class will need to modified as well if any
additional changes are made to shell script.